Trekker ut Visning og Endring av vilkårsvurdering i egen fil
